Update add_timestamp.py
Added basic commandLine interface

+++ AutomatedUtilities/add_timestamp.py
@@ -6,29 +6,39 @@ from automated_utilities import AutomatedUtilities
 
 print(f"Running... {sys.argv[0]}")
 
-au = AutomatedUtilities()
+while(True):
+    print("1.Run program")
+    print("0.Exit")
+    choice=input("Your choice number")
+    if int(choice)==1:
+        au = AutomatedUtilities()
+        parent_folder=input("Please input parent folder:")
+        au.setOrigin(parent_folder+"\\")
+        print(f"Parameter 1: {au.getOrigin()}")
+        destination_folder=input("Please input destination folder:")
+        au.setDestination(destination_folder+"\\")
+        print(f"Parameter 2: {au.getDestination()}")
+        the_file=input("Please give the name of the file without its type")
+        au.setFilename(the_file)
+        print(f"Parameter 3: {au.getFilename()}")
+        # set origin  as the current working directory
+        os.chdir(au.getOrigin())
 
-au.setOrigin(sys.argv[1])
-print(f"Parameter 1: {au.getOrigin()}")
+        timestamp = au.getTimestamp()
 
-au.setDestination(sys.argv[2])
-print(f"Parameter 2: {au.getDestination()}")
+        origin = au.getOrigin()
+        destination = au.getDestination()
+        filename = au.getFilename()
 
-au.setFilename(sys.argv[3])
-print(f"Parameter 3: {au.getFilename()}")
+        newfile = au.copyFile(origin, filename)
+        newfile = au.removeExtentions(newfile)
+        new_filename = au.addTimestamp(origin, newfile)
+        zippedfile = au.zipFile(origin, destination, new_filename)
+        au.moveFile(zippedfile, destination)
+        au.deleteFile(new_filename, origin)
+    elif int(choice)==0:
+        print("Program will now exit!")
+        exit()
 
-# set origin  as the current working directory
-os.chdir(au.getOrigin())
 
-timestamp = au.getTimestamp()
 
-origin = au.getOrigin()
-destination = au.getDestination()
-filename = au.getFilename()
-
-newfile = au.copyFile(origin, filename)
-newfile = au.removeExtentions(newfile)
-new_filename = au.addTimestamp(origin, newfile)
-zippedfile = au.zipFile(origin, destination, new_filename)
-au.moveFile(zippedfile, destination)
-au.deleteFile(new_filename, origin)
